requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.7
Django==3.2.6
django-allauth==0.42.0
django-cors-headers==3.11.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.4.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tldextract==1.6
toml==0.10.2
unicodecsv==0.13.0
urllib3==1.26.8
whitenoise==5.3.0